refactor: extract remote MCP startup
